backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.6.0
google-api-python-client==2.146.0
google-cloud-firestore==2.19.0
python-dateutil==2.9.0.post0
rapidfuzz==3.9.7
pytesseract==0.3.10
Pillow==10.4.0
